% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleomap}
\alias{paleomap}
\title{PALEOMAP reconstruction file}
\format{
A data frame with 196 variables:
\describe{
\item{lng}{Reference longitude for the 1° x 1° spatial grid}
\item{lat}{Reference latitude for the 1° x 1° spatial grid}
\item{lng_1}{Palaeolongitude of the reference longitude at 1 Ma}
\item{lat_1}{Palaeolatitude of the reference latitude at 1 Ma}
...
}
}
\usage{
paleomap
}
\description{
Reconstruction files provide pre-generated palaeocoordinates for a
1° x 1° spatial grid which can be used to spatiotemporally link
fossil localities (or other geographic data) with their position in the
geological past. Rotations were generated for the midpoint age of all
Phanerozoic stratigraphic stages (0--540 Ma).
\cr
\cr
This file provides palaeorotations from the PALEOMAP plate rotation model
(Scotese & Wright, 2018).
Palaeocoordinates were generated via the GPlates API service, available at
the following: \url{https://gwsdoc.gplates.org}. This model has a temporal
coverage of 0--750 Ma.
\cr
\cr
\strong{NOTE}: two modifications have been made to the data accessed via the API
service:
(1) data has been rounded to two decimal places to reduce file size, and
(2) palaeocoordinates which do not change across the whole time series are
assumed to be areas not covered by the models, and are updated to
NA accordingly.
}
\section{Reference}{

Scotese, C., & Wright, N. M. (2018). PALEOMAP Paleodigital Elevation Models
(PaleoDEMs) for the
Phanerozoic. PALEOMAP Project.
\url{https://www.earthbyte.org/paleodem-resource-scotese-and-wright-2018/}.

See GPlates documentation for additional information and details:
\url{https://gwsdoc.gplates.org/reconstruction}.
}

\keyword{datasets}
