% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uncertainty}
\alias{uncertainty}
\title{Uncertainty reconstruction file}
\format{
A data frame with 196 variables:
\describe{
\item{lng}{Reference longitude for the 1° x 1° spatial grid}
\item{lat}{Reference latitude for the 1° x 1° spatial grid}
\item{lng_1}{Palaeolongitudinal range of the reference longitude at 1 Ma
for all reconstruction files}
\item{lat_1}{Palaeolatitudinal range of the reference longitude at 1 Ma
for all reconstruction files}
...
}
}
\usage{
uncertainty
}
\description{
This file provides the range in reconstructed palaeocoordinates from all
available reconstruction files for each cell in the 1° x 1° spatial
grid.
}
\keyword{datasets}
